src/requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
flake8==6.1.0
greenlet==3.0.1
h11==0.14.0
idna==3.6
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
